+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -397,6 +397,7 @@ sub build_isf_public_interface_contract {
             't/1243-isf-port-binding-schedule-report.t',
             't/1244-isf-wait-clause-lowering.t',
             't/1245-isf-transaction-loop-lowering.t',
+            't/1246-isf-setter-syntax.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',